build: Terminado endpoints, empezado register front

+++ confiteria-santana-back/src/main/java/com/vortex/infrastructure/repositories/PaymentsDAO.java
@@ -34,7 +34,7 @@ public class PaymentsDAO {
      * @param id the id
      * @return the payments
      */
-    public Payments find(int id) {
+    public Payments find(Long id) {
         return em.find(Payments.class, id);
     }
